+++ lib/approc_image_picker.dart
@@ -1,3 +1,4 @@
 export 'src/prepare_jpeg.dart';
 export 'src/prepare_sheet.dart';
+export 'src/prepare_theme.dart';
 export 'src/picker.dart';